Refactor to pave way for using new LDBWS API.
This commit doesn't change any of the features of the library yet. It
simply moves each piece of functionality into it's own file, since the
code base is about to get much larger with the upgrade to the latest
version of the LDBWS API and the various new requests and responses that
form part of it.

This is a gently backwards incompatible change, taking the opportunity
to rename the module based on the fact that the Darwin brand will soon
be getting a new Web Service as part of it which is an entirely separate
thing. Thus using 'ldbws' for the module name for the LDBWS web service
makes more sense than just calling it 'webservice' once there are two
separate web services supported by this library.